--- src/components/ProtectedRoute.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { Navigate } from "react-router-dom";
-import { useUserAuth } from "../context/userAuthContext";
-const ProtectedRoute = ({ children }) => {
-  const { user } = useUserAuth();
-  if (!user) {
-    return <Navigate to="/" />;
-  }
-  return children;
-};
-
-export default ProtectedRoute;
