+++ src/components/ProtectedRoute.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Navigate } from "react-router-dom";
+import { useUserAuth } from "../context/userAuthContext";
+const ProtectedRoute = ({ children }) => {
+  const { user } = useUserAuth();
+  if (!user) {
+    return <Navigate to="/" />;
+  }
+  return children;
+};
+
+export default ProtectedRoute;
